+++ aat-lib/src/main/java/ch/bailu/aat_lib/gpx/GpxDataWrapper.java
@@ -62,7 +62,7 @@ public class GpxDataWrapper implements GpxDeltaPointInterface, GpxBigDeltaInterf
     }
 
     @Override
-    public double getAltitude() {
+    public float getAltitude() {
         return point.getAltitude();
     }
